feat: 重构SDK初始化逻辑并修复导入错误
将SDK初始化方法从sdk.py移动到__init__.py，修复直接从ErisPulse导入时的错误
将CLI中的reloader直接移动到run命令内部定义
移除模块管理器中多余的警告逻辑
