+++ CHANGELOG.md
@@ -32,6 +32,15 @@
 ```
 
 ---
+## [2.3.4-dev.3] - 2026/02/04
+> 开发版本
+
+### 变更
+- @wsu2059q
+  - 修复入口直接从ErisPulse导入时发生错误的问题
+  - CLI中的reloader直接移动到run命令内部进行定义
+
+---
 
 ## [2.3.4-dev.2] - 2026/02/03
 > 开发版本
