+++ src/ErisPulse/CLI/utils/__init__.py
@@ -5,9 +5,7 @@ ErisPulse SDK 工具模块
 """
 
 from .package_manager import PackageManager
-from .reload_handler import ReloadHandler
 
 __all__ = [
-    "PackageManager",
-    "ReloadHandler"
+    "PackageManager"
 ]
